feat(types): Enable types when importing from /src Previously, when using the 'Advanced Installation' method outlined in docs, types would be unavailable. By relocating the component .d.ts files to @types/src, and pointing the index.d.ts imports to these, types become available for users importing both the compiled package and those importing from the package source.
