+++ packages/svelte-materialify/@types/index.d.ts
@@ -1,50 +1,50 @@
-export { default as ClickOutside } from './ClickOutside';
-export { default as Intersect } from './Intersect';
-export { default as Jump } from './Jump';
-export { default as Ripple } from './Ripple';
-export { default as Touch } from './Touch';
+export { default as ClickOutside } from './src/ClickOutside';
+export { default as Intersect } from './src/Intersect';
+export { default as Jump } from './src/Jump';
+export { default as Ripple } from './src/Ripple';
+export { default as Touch } from './src/Touch';
 
-export { default as Alert } from './Alert';
-export { default as AppBar } from './AppBar';
-export { default as Avatar } from './Avatar';
-export { default as Badge } from './Badge';
-export { default as Breadcrumbs } from './Breadcrumbs';
-export { default as Button } from './Button';
-export { default as Card } from './Card';
-export { default as CardActions } from './CardActions';
-export { default as CardSubtitle } from './CardSubtitle';
-export { default as CardText } from './CardText';
-export { default as CardTitle } from './CardTitle';
-export { default as Checkbox } from './Checkbox';
-export { default as Col } from './Col';
-export { default as Container } from './Container';
-export { default as Dialog } from './Dialog';
-export { default as Divider } from './Divider';
-export { default as ExpansionPanel } from './ExpansionPanel';
-export { default as ExpansionPanels } from './ExpansionPanels';
-export { default as Icon } from './Icon';
-export { default as Input } from './Input';
-export { default as ItemGroup } from './ItemGroup';
-export { default as List } from './List';
-export { default as ListGroup } from './ListGroup';
-export { default as ListItem } from './ListItem';
-export { default as ListItemGroup } from './ListItemGroup';
-export { default as MaterialApp } from './MaterialApp';
-export { default as Menu } from './Menu';
-export { default as NavigationDrawer } from './NavigationDrawer';
-export { default as Overlay } from './Overlay';
-export { default as ProgressCircular } from './ProgressCircular';
-export { default as ProgressLinear } from './ProgressLinear';
-export { default as Radio } from './Radio';
-export { default as Row } from './Row';
-export { default as Select } from './Select';
-export { default as Slider } from './Slider';
-export { default as Snackbar } from './Snackbar';
-export { default as Subheader } from './Subheader';
-export { default as Switch } from './Switch';
-export { default as Tab } from './Tab';
-export { default as Tabs } from './Tabs';
-export { default as TabContent } from './TabContent';
-export { default as Textarea } from './Textarea';
-export { default as TextField } from './TextField';
-export { default as Tooltip } from './Tooltip';
+export { default as Alert } from './src/Alert';
+export { default as AppBar } from './src/AppBar';
+export { default as Avatar } from './src/Avatar';
+export { default as Badge } from './src/Badge';
+export { default as Breadcrumbs } from './src/Breadcrumbs';
+export { default as Button } from './src/Button';
+export { default as Card } from './src/Card';
+export { default as CardActions } from './src/CardActions';
+export { default as CardSubtitle } from './src/CardSubtitle';
+export { default as CardText } from './src/CardText';
+export { default as CardTitle } from './src/CardTitle';
+export { default as Checkbox } from './src/Checkbox';
+export { default as Col } from './src/Col';
+export { default as Container } from './src/Container';
+export { default as Dialog } from './src/Dialog';
+export { default as Divider } from './src/Divider';
+export { default as ExpansionPanel } from './src/ExpansionPanel';
+export { default as ExpansionPanels } from './src/ExpansionPanels';
+export { default as Icon } from './src/Icon';
+export { default as Input } from './src/Input';
+export { default as ItemGroup } from './src/ItemGroup';
+export { default as List } from './src/List';
+export { default as ListGroup } from './src/ListGroup';
+export { default as ListItem } from './src/ListItem';
+export { default as ListItemGroup } from './src/ListItemGroup';
+export { default as MaterialApp } from './src/MaterialApp';
+export { default as Menu } from './src/Menu';
+export { default as NavigationDrawer } from './src/NavigationDrawer';
+export { default as Overlay } from './src/Overlay';
+export { default as ProgressCircular } from './src/ProgressCircular';
+export { default as ProgressLinear } from './src/ProgressLinear';
+export { default as Radio } from './src/Radio';
+export { default as Row } from './src/Row';
+export { default as Select } from './src/Select';
+export { default as Slider } from './src/Slider';
+export { default as Snackbar } from './src/Snackbar';
+export { default as Subheader } from './src/Subheader';
+export { default as Switch } from './src/Switch';
+export { default as Tab } from './src/Tab';
+export { default as Tabs } from './src/Tabs';
+export { default as TabContent } from './src/TabContent';
+export { default as Textarea } from './src/Textarea';
+export { default as TextField } from './src/TextField';
+export { default as Tooltip } from './src/Tooltip';
