+++ packages/svelte-materialify/@types/src/Avatar.d.ts
