+++ packages/svelte-materialify/@types/src/Badge.d.ts
