+++ packages/svelte-materialify/@types/src/Card.d.ts
