+++ packages/svelte-materialify/@types/src/CardSubtitle.d.ts
