+++ packages/svelte-materialify/@types/src/CardText.d.ts
