+++ packages/svelte-materialify/@types/src/ClickOutside.d.ts
