+++ packages/svelte-materialify/@types/src/Dialog.d.ts
