+++ packages/svelte-materialify/@types/src/Divider.d.ts
