+++ packages/svelte-materialify/@types/src/ExpansionPanel.d.ts
