+++ packages/svelte-materialify/@types/src/ExpansionPanels.d.ts
