+++ packages/svelte-materialify/@types/src/Icon.d.ts
