+++ packages/svelte-materialify/@types/src/Intersect.d.ts
