+++ packages/svelte-materialify/@types/src/MaterialApp.d.ts
