+++ packages/svelte-materialify/@types/src/MaterialAppMin.d.ts
