+++ packages/svelte-materialify/@types/src/NavigationDrawer.d.ts
