+++ packages/svelte-materialify/@types/src/ProgressCircular.d.ts
