+++ packages/svelte-materialify/@types/src/Ripple.d.ts
