+++ packages/svelte-materialify/@types/src/Slider.d.ts
