+++ packages/svelte-materialify/@types/src/Snackbar.d.ts
