+++ packages/svelte-materialify/@types/src/Switch.d.ts
