+++ packages/svelte-materialify/@types/src/Tab.d.ts
