+++ packages/svelte-materialify/@types/src/TabContent.d.ts
