+++ packages/svelte-materialify/@types/src/Tabs.d.ts
