+++ packages/svelte-materialify/@types/src/shared.d.ts
